Fix: int to long

+++ HotSource/src/main/java/hotsource/model/asset/AssetService.java
@@ -6,10 +6,10 @@ import hotsource.domain.Asset;
 
 public interface AssetService {
 	public List selectAll();
-	public Asset select(int asset_id);
-	public List selectBySellerId(int seller_id);
+	public Asset select(long asset_id);
+	public List selectBySellerId(long seller_id);
 	
 	public void regist(Asset asset);
 	public void update(Asset asset);
-	public void delete(int asset_id);
+	public void delete(long asset_id);
 }
