+++ HotSource/src/main/java/hotsource/model/notice_like/NoticeLikeDAO.java
@@ -12,5 +12,5 @@ import hotsource.domain.NoticeLike;
 public interface NoticeLikeDAO {
 
 	public List selectAll();
-	public NoticeLike select(int notice_like_id);
+	public NoticeLike select(long notice_like_id);
 }
