+++ HotSource/src/main/java/hotsource/model/review/ReviewDAO.java
@@ -6,8 +6,8 @@ import hotsource.domain.Review;
 
 public interface ReviewDAO {
 	public List selectAll();
-	public Review select(int review_id);
-	public List selectByUserId(int user_id);
-	public List selectByAssetId(int asset_id);
-	public Review selectByOrderedAssetId(int order_asset_id);
+	public Review select(long review_id);
+	public List selectByUserId(long user_id);
+	public List selectByAssetId(long asset_id);
+	public Review selectByOrderedAssetId(long order_asset_id);
 }
