+++ HotSource/src/main/java/hotsource/model/subscribe/SubscribeDAO.java
@@ -4,6 +4,6 @@ import java.util.List;
 
 public interface SubscribeDAO {
 	public List selectAll();
-	public List selectByUserId(int user_id);
-	public List selectBySellerId(int seller_id);
+	public List selectByUserId(long user_id);
+	public List selectBySellerId(long seller_id);
 }
