+++ HotSource/src/main/java/hotsource/model/user/UserDAO.java
@@ -6,7 +6,7 @@ import hotsource.domain.User;
 
 public interface UserDAO {
 	public List selectAll();
-	public User select(int user_id);
-	public List selectByRoleId(int role_id);
+	public User select(long user_id);
+	public List selectByRoleId(long role_id);
 	public void insert(User user);
 }
